Ignore the time slice signal in enter_syscall()
This seems to solve the issue with very occational failure
in the clone_interruption test during record. I've have noticed
the failure in -32 version of these tests but my investigation
has not uncovered any reason why the same failure mode may
not apply to x64.

The SIGSTKFLT signal (that is more frequent due to the -c100) in
this test can sometimes interrupt the clone syscall in prepare_clone() and
subsequently cause the signal to be stashed in enter_syscall() and
that would change the stashed_signals_blocking_more_signals to true.
This eventually leads to the the signal mask to be changed
and when a new thread would be created, it would inherit
this new mask. Unfortunately in this new thread, the signal
mask would not get reset. Later, when kill would try to send
a SIGCHLD to the newly created thread, the signal would not
reach due to this aforementioned mask that only allowed desched
and time slice.